requirements.txt
appdirs==1.4.4
appnope==0.1.2
astroid==2.5.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
geocoder==1.38.1
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
pylint==2.7.3
python-dotenv==0.16.0
ratelim==0.1.6
requests==2.25.1
six==1.15.0
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
virtualenv==20.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
